# server-src/swagger.json
{
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "definitions": {
    "ITeamVm": {
      "properties": {
        "teamName": {
          "type": "string"
        },
        "teamId": {
          "type": "string"
        },
        "isInitialized": {
          "type": "boolean"
        },
        "_id": {
          "type": "string"
        }
      },
      "type": "object"
    },
    "ITicketResponse": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "summary": {
          "type": "string"
        },
        "ticketNumber": {
          "type": "number",
          "format": "double"
        },
        "slug": {
          "type": "string"
        },
        "createdOn": {
          "type": "string",
          "format": "date-time"
        },
        "isResolved": {
          "type": "boolean"
        },
        "team": {
          "$ref": "#/definitions/ITeamVm"
        }
      },
      "type": "object"
    },
    "ISystemAuthResponse": {
      "properties": {
        "authToken": {
          "type": "string"
        }
      },
      "required": [
        "authToken"
      ],
      "type": "object"
    },
    "IAuthLoginParams": {
      "properties": {
        "nick": {
          "type": "string"
        },
        "pass": {
          "type": "string"
        }
      },
      "required": [
        "nick",
        "pass"
      ],
      "type": "object"
    }
  },
  "info": {
    "description": "Documentation of APIs utilizing data gathered from uTicket Slack bot",
    "version": "1.0.0",
    "title": "uTicket Slack Bot"
  },
  "paths": {
    "/teams": {
      "get": {
        "operationId": "GetTeams",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ITeamVm"
              }
            }
          }
        },
        "tags": [
          "Operations"
        ],
        "security": [],
        "parameters": []
      }
    },
    "/tickets": {
      "get": {
        "operationId": "GetTickets",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ITicketResponse"
              }
            }
          }
        },
        "tags": [
          "Operations"
        ],
        "security": [],
        "parameters": []
      }
    },
    "/tickets/{teamName}": {
      "get": {
        "operationId": "GetTicketsByTeam",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ITicketResponse"
              }
            }
          }
        },
        "tags": [
          "Operations"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "teamName",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/auth/auth": {
      "post": {
        "operationId": "AdminAuth",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/ISystemAuthResponse"
            }
          }
        },
        "tags": [
          "System"
        ],
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "loginParams",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IAuthLoginParams"
            }
          }
        ]
      }
    },
    "/auth/clear/{collection}": {
      "get": {
        "operationId": "ClearDatabase",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "boolean"
            }
          }
        },
        "tags": [
          "System"
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "collection",
            "required": true,
            "type": "string"
          }
        ]
      }
    }
  },
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "tags": [
    {
      "name": "Operations",
      "description": "Application main operations"
    },
    {
      "name": "System",
      "description": "Admin operations"
    }
  ]
}